+++ src/main/java/info/henrysson/advent_bonus/service/MatProductService.java
@@ -6,5 +6,5 @@ import java.io.IOException;
 import java.util.List;
 
 public interface MatProductService {
-    List<MatProduct> getCategoryProducts(int categoryId) throws IOException;
+    List<MatProduct> getCategoryProducts(long categoryId) throws IOException;
 }
